Add files via upload
The telecommunications operator wants to learn how to predict customer churn. If it turns out that a user is planning to leave, they will be offered promotional codes and special conditions. The operator's team has gathered personal data about some clients, information about their tariffs, and contracts.
The goal of the project is to train a model that will predict the possibility of a customer leaving the company. The model should achieve an AUC-ROC score above 0.85.